test: fix ios browserstack appium set up
