+++ .github/workflows/ui-tests.yml
@@ -141,35 +141,6 @@ jobs:
       - name: Build iOS app
         working-directory: sample
         run: ./build_ios.sh
-      - name: Install jq (if not installed)
-        run: |
-          if ! command -v jq &> /dev/null; then
-            echo "jq not found, installing..."
-            brew install jq
-          else
-            echo "jq is already installed"
-          fi
-      - name: Upload iOS app to BrowserStack
-        working-directory: sample/Tests
-        env:
-          BROWSERSTACK_USERNAME: ${{ secrets.BROWSERSTACK_USERNAME }}
-          BROWSERSTACK_ACCESS_KEY: ${{ secrets.BROWSERSTACK_ACCESS_KEY }}
-        run: |
-          curl -u "${{ secrets.BROWSERSTACK_USERNAME }}:${{ secrets.BROWSERSTACK_ACCESS_KEY }}" \
-          -X POST "https://api-cloud.browserstack.com/app-automate/upload" \
-          -F "file=@Payload.ipa"
-
-          # Extract app_url from the response using jq
-          app_url=$(echo "$response" | jq -r '.app_url')
-
-          if [ -z "$app_url" ]; then
-            echo "Error: app_url is not found in the response"
-            exit 1  # Fail the workflow if app_url is not present
-          fi
-
-          echo "app_url: $app_url"
-
-          echo "APP_URL=$app_url" >> $GITHUB_ENV
       - uses: actions/setup-python@v4
         with:
           python-version: "3.10"
@@ -180,6 +151,6 @@ jobs:
           MAILSLURP_API_KEY: ${{ secrets.MAILSLURP_API_KEY }}
           BROWSERSTACK_USERNAME: ${{ secrets.BROWSERSTACK_USERNAME }}
           BROWSERSTACK_ACCESS_KEY: ${{ secrets.BROWSERSTACK_ACCESS_KEY }}
-        working-directory: sample/Tests
-        run: browserstack-sdk pytest -s ./test/test_ios.py --browserstack.config "browserstack.ios.yml"
+        working-directory: sample/Tests/test/ios
+        run: browserstack-sdk pytest -xs ./test_ios.py --browserstack.config "browserstack.ios.yml"
