+++ sample/ProjectSettings/ProjectSettings.asset
@@ -157,13 +157,13 @@ PlayerSettings:
   applicationIdentifier:
     Android: com.immutable.ImmutableSample
     Standalone: com.immutable.Immutable-Sample
-    iPhone: com.immutable.Immutable-Sample
+    iPhone: com.immutable.Immutable-Sample-GameSDK
   buildNumber:
     Standalone: 0
     VisionOS: 0
     iPhone: 0
     tvOS: 0
-  overrideDefaultApplicationIdentifier: 0
+  overrideDefaultApplicationIdentifier: 1
   AndroidBundleVersionCode: 1
   AndroidMinSdkVersion: 26
   AndroidTargetSdkVersion: 33
