+++ sample/Tests/test/ios/browserstack.ios.yml
@@ -21,7 +21,7 @@ buildName: iOS build
 # Read more about buildIdentifiers here -> https://www.browserstack.com/docs/automate/selenium/organize-tests
 buildIdentifier: '#${BUILD_NUMBER}' # Supports strings along with either/both ${expression}
 framework: pytest
-source: pytest-browserstack:sample-sdk:v1.0
+source: pytest-appium-browserstack:sample-sdk:v1.0
 
 # Set `app` to define the app that is to be used for testing.
 # It can either take the id of any uploaded app or the path of the app directly.
@@ -35,9 +35,9 @@ app: ./Payload.ipa #For running local tests
 # Entire list available here -> (https://www.browserstack.com/list-of-browsers-and-platforms/automate)
 
 platforms:
-  - platformName: ios
-    deviceName: iPhone 14 Pro Max
-    platformVersion: 16
+  - deviceName: iPhone 14 Pro
+    osVersion: 16
+    platformName: ios
 
 # =======================
 # Parallels per Platform
@@ -57,10 +57,10 @@ parallelsPerPlatform: 1
 # Set browserStackLocal to true if your website under test is not accessible publicly over the internet
 # Learn more about how BrowserStack Local works here -> https://www.browserstack.com/docs/automate/selenium/local-testing-introduction
 browserstackLocal: true # <boolean> (Default false)
-browserStackLocalOptions:
+# browserStackLocalOptions:
   #Options to be passed to BrowserStack local in-case of advanced configurations
   #  localIdentifier: # <string> (Default: null) Needed if you need to run multiple instances of local.
-  forceLocal: true  # <boolean> (Default: false) Set to true if you need to resolve all your traffic via BrowserStack Local tunnel.
+  # forceLocal: true  # <boolean> (Default: false) Set to true if you need to resolve all your traffic via BrowserStack Local tunnel.
 # Entire list of arguments available here -> https://www.browserstack.com/docs/automate/selenium/manage-incoming-connections
 
 # ===================
@@ -70,4 +70,3 @@ debug: false # <boolean> # Set to true if you need screenshots for every seleniu
 networkLogs: false # <boolean> Set to true to enable HAR logs capturing
 consoleLogs: errors # <string> Remote browser's console debug levels to be printed (Default: errors)
 # Available options are `disable`, `errors`, `warnings`, `info`, `verbose` (Default: errors)
-acceptInsecureCerts: true
