+++ sample/Tests/test/test.py
@@ -11,7 +11,7 @@ class TestConfig:
     PASSPORT_ID="email|67480492219c150aceeb1f37"
     WALLET_ADDRESS = "0x547044ea95f03651139081241c99ffedbefdc5e8"
     ANDROID_PACKAGE = "com.immutable.ImmutableSample"
-    IOS_BUNDLE_ID = "com.immutable.Immutable-Sample"
+    IOS_BUNDLE_ID = "com.immutable.Immutable-Sample-GameSDK"
     
 class UnityTest(unittest.TestCase):
